+++ src/main/java/io/wkna/sdp/messages/SyncTickMessage.java
@@ -15,7 +15,7 @@ public class SyncTickMessage extends DemoMessage {
     }
 
     @Override
-    public int writeData(ByteBuffer dst, int remaining, int offset) {
+    public int writeData(ByteBuffer dst, int remaining, long offset) {
         return 0;
     }
